% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abb_6.6}
\alias{abb_6.6}
\title{BT24 default plot settings.}
\format{
\subsection{\code{abb_6.6}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
abb_6.6
}
\description{
Default plot settings Abbildung 6.6 from BT22.
}
\keyword{datasets}
