% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abb_8.4}
\alias{abb_8.4}
\title{BT24 default plot settings.}
\format{
\subsection{\code{abb_8.4}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
abb_8.4
}
\description{
Default plot settings Abbildung 8.4 from BT22.
}
\keyword{datasets}
