% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abb_9.1}
\alias{abb_9.1}
\title{BT24 default plot settings.}
\format{
\subsection{\code{abb_9.1}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
abb_9.1
}
\description{
Default plot settings Abbildung 9.1 from BT22 (stacked barplot).
}
\keyword{datasets}
