% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjusted_means}
\alias{adjusted_means}
\title{BT21 adusted means}
\format{
\subsection{\code{adjusted_means}}{

A data frame with 264 rows and 12 columns:
\describe{
\item{est}{Estimated value.}
\item{p}{p-value}
...
}
}
}
\source{
<"Q:/BT2021/BT/60_Bericht/04_Mittelwerte/adjustierte Mittelwerte/Ergebnisse/01_ohneTrend_z_standard.xlsx">
}
\usage{
adjusted_means
}
\description{
BT21 data set that contains the results of the adjusted means chapter.
}
\keyword{datasets}
