% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjusted_means_diff}
\alias{adjusted_means_diff}
\title{BT21 adusted mean differences}
\format{
\subsection{\code{adjusted_means_diff}}{

A data frame with 128 rows and 12 columns:
\describe{
\item{est}{Mean difference}
\item{p}{p-value}
...
}
}
}
\source{
<"Q:/BT2021/BT/60_Bericht/04_Mittelwerte/adjustierte Mittelwerte/Ergebnisse/01_ohneTrend_z_standard.xlsx">
}
\usage{
adjusted_means_diff
}
\description{
A randomly permuted data set stemming from the BT21 ajdusted means data file displaying the (randomly permuted!) the differnce of the countrys adjusted and unadjusted means from the german mean value.
}
\keyword{datasets}
