% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anteile}
\alias{anteile}
\title{BT24 example data.}
\format{
\subsection{\code{anteile}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
anteile
}
\description{
Example data for stacked bar charts.
}
\keyword{datasets}
