% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_MinSta}
\alias{barplot_MinSta}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_MinSta}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_MinSta
}
\description{
Default plot settings for barplots.
}
\keyword{datasets}
