% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_MinSta_trend}
\alias{barplot_MinSta_trend}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_MinSta_trend}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_MinSta_trend
}
\description{
Default plot settings for Mindest Standards trend.
}
\keyword{datasets}
