% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_noTrend}
\alias{barplot_noTrend}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_noTrend}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_noTrend
}
\description{
Default plot settings for mean-barplot without trend.
}
\keyword{datasets}
