% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_plot_frame}
\alias{barplot_plot_frame}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_plot_frame}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_plot_frame
}
\description{
Default plot settings for Mindest Standard.
}
\keyword{datasets}
