% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_plot_frz}
\alias{barplot_plot_frz}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_plot_frz}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_plot_frz
}
\description{
Default plot settings for Mindest Standard Französisch.
}
\keyword{datasets}
