% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barplot_table_plot_pattern}
\alias{barplot_table_plot_pattern}
\title{BT22 default plot settings.}
\format{
\subsection{\code{barplot_table_plot_pattern}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}

\subsection{\code{barplot_table_plot_pattern}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
barplot_table_plot_pattern

barplot_table_plot_pattern
}
\description{
Default plot settings for Mindest Standards trend.

Default plot settings for Mindest Standards trend.
}
\keyword{datasets}
