% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_plot_borders.R
\name{calc_plot_borders}
\alias{calc_plot_borders}
\title{Get min and max of a vector and round them to 10.}
\usage{
calc_plot_borders(x, accuracy = 10)
}
\arguments{
\item{x}{Numeric vector.}

\item{accuracy}{Number to round to.}
}
\value{
Negative and positive plot borders.
}
\description{
Currently only works, if the plot center is zero. Min and Max values of the input variable are calculated. The higher of these two values is put out as negative and positive value.
}
\examples{
calc_plot_borders(c(11, -14))
}
