% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lineplot.R
\name{calc_plot_lims}
\alias{calc_plot_lims}
\title{Calculate different plot limit values.}
\usage{
calc_plot_lims(plot_dat, point_values, plot_settings)
}
\arguments{
\item{plot_dat}{Input is a list prepared by \code{\link[=prep_plot]{prep_plot()}}.`}

\item{point_values}{Character string of the column name in \code{plot_dat[["plot_points"]]} containing the y-values for the plotted points. Defaults to \code{est_noTrend_noComp}.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
List containing the following elements:
\itemize{
\item \code{range_y}: Minimum and maximum of the values in \code{point_values}.
\item \code{y_lims_total}: Minimum and maximum value of the plot.
\item \code{coords}: Y-value of the first brace start, and heighest y-value of the plot.
}
}
\description{
Calculate different plot limit values.
}
\examples{
#tbd
}
