% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sig.R
\name{calc_sig}
\alias{calc_sig}
\title{Compute if a p value is significant.}
\usage{
calc_sig(data, sig_niveau, p_column = "p")
}
\arguments{
\item{data}{Input data.frame.}

\item{sig_niveau}{Significance niveau.}

\item{p_column}{Column in data which contains the p-values.}
}
\value{
The data.frame with a new column named 'sig'
}
\description{
Helper function.
}
