% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Fill in and rename necessary columns}
\usage{
clean_data(data, grouping_var, BLs, groups, competence)
}
\arguments{
\item{data}{Input data.frame stemming from eatRep.}

\item{grouping_var}{Character string containing the column in \code{data} that should be used to distinguish between subgroups.}

\item{BLs}{bundesländer}

\item{groups}{grouping_var groups}

\item{competence}{Character string containing the competence that should be plotted. Currently Has to be found in \code{data$kb} (even though that should be made optional in the future).}
}
\value{
\code{clean_data()} returns a \code{data} with renamend columns and filled in \code{NAs}.
}
\description{
\code{clean_data()} performs some simple data wrangling to prepare the input \code{data.frame} for the more plot specific data transformations.
}
\details{
The following operations are performed on \code{data}:
\itemize{
\item For consistency, points in column names are subbed with underscores.
\item Only rows containing \code{mean} in \code{data$parameter} are returend.
\item The group \code{wholeGroup} currently contains \code{NAs} in the columns \code{data$TR_BUNDESLAND} and \code{data$grouping_var}. These are subbed with \code{wholeGroup} or \code{noGroup} respectively.
\item The column \code{data$TR_BUNDESLAND} is filled up by extracting the first state found in \code{data$groups} for the respective row.
}
}
\examples{
# tbd
}
