% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmyk.R
\name{cmyk}
\alias{cmyk}
\title{Set colours as cmyk values.}
\usage{
cmyk(C, M, Y, K)
}
\arguments{
\item{C}{Numeric cyan value.}

\item{M}{Numeric magenta value.}

\item{Y}{Numeric yellow value.}

\item{K}{Numeric key (black) value.}
}
\value{
RGB colour value.
}
\description{
Set colours as cmyk values.
}
\examples{
cmyk(0, 0, 0, 1)
}
