% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine different ggplots}
\usage{
combine_plots(plot_list, plot_widths = NULL)
}
\arguments{
\item{plot_list}{List of the plots that should be combined.}

\item{plot_widths}{Set the widths of the plots manually. Defaults to \code{NULL}, in which case the width will be calculated so the x axis scales are on the same metric.}
}
\value{
Patchwork plot.
}
\description{
Combine different ggplots
}
\examples{
# tbd
}
