% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{connect_points}
\alias{connect_points}
\title{Connect Points}
\usage{
connect_points(data_prep, grouping_var)
}
\arguments{
\item{grouping_var}{Variable for grouping}

\item{my_data}{Data}
}
\value{
ggplot object
}
\description{
Connect Points
}
