% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_directional_sig.R
\name{construct_directional_sig}
\alias{construct_directional_sig}
\title{Calculate the significance descriptions for the 'Mindeststandards'}
\usage{
construct_directional_sig(dat_bar, est_column, sig_column)
}
\arguments{
\item{dat_bar}{Data prepared by \code{prep_plot()}.}

\item{est_column}{Column that contains the estimates.}

\item{sig_column}{Column that contains the significances.}
}
\value{
Dataframe with an added column with a name that ends with \verb{_directional_sig}.
}
\description{
Calculate the significance descriptions for the 'Mindeststandards'
}
\examples{
#tbd
}
