% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_grouping_var.R
\name{construct_grouping_var}
\alias{construct_grouping_var}
\title{Merge two grouping var columns into one.}
\usage{
construct_grouping_var(dat, grouping_vars, group_var = "group")
}
\arguments{
\item{grouping_vars}{}
}
\description{
Merge two grouping var columns into one.
}
