% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_label.R
\name{construct_label_2}
\alias{construct_label_2}
\title{Construct labels}
\usage{
construct_label_2(
  dat,
  new_name = "label",
  label_est = NULL,
  label_se = NULL,
  label_sig_bold = NULL,
  label_sig_superscript = NULL,
  label_sig_superscript_extra_column = FALSE,
  round_est = 0,
  round_se = 1,
  plot_settings = plotsettings_tablebarplot()
)
}
\arguments{
\item{dat}{Data frame with the columns that should be merged into labels.}

\item{new_name}{Character string for the new column that is added to \code{dat}. Defaults to \code{'label'}.}

\item{label_est}{Character string for the column with the estimates.}

\item{label_se}{Character string for the column with the standard errors.}

\item{label_sig_bold}{Character string for the column with the significant estimates that should determine the bold printing.}

\item{label_sig_superscript}{Character string for the column with the significant estimates that should determine the superscript.}

\item{label_sig_superscript_extra_column}{Logical, if set 'FALSE' the superscript for significant values is added directly into the label (necessary for line plots), if set 'TRUE' the superscript for significant values is written into an extra column with the ending '_sig_superscript' (necessary for tables).}

\item{round_est}{Rounding of label_est.}

\item{round_se}{Rounding of label_se.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
The data frame with an added column for the constructed label.
}
\description{
This function creates a new column or label, merging estimates and standard errors with significant estimates represented in bold or as superscript (via \code{label_est}), and standard errors in brackets, if needed (via \code{label_se}). NAs are converted to empty strings. Main usage is for plotting tables and brace labels.
}
\examples{
# example data frame
dat <- data.frame(
  names = c("Berlin", "Hamburg", "Hessen", "Niedersachsen", "Saarland"),
  estimate = c(400, 650, 380, 500, 600),
  se = c(0.1, 0.45, 1, 0.27, 0.9),
  p_estimate = c(FALSE, FALSE, TRUE, TRUE, FALSE)
)
}
