% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dispar}
\alias{dispar}
\title{BT24 example data.}
\format{
\subsection{\code{dispar}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
dispar
}
\description{
Example data with multiple competence areas.
}
\keyword{datasets}
