% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_dat.R
\name{filter_plot_dat}
\alias{filter_plot_dat}
\title{Filter specified rows from plot_dat objects.}
\usage{
filter_plot_dat(
  plot_dat,
  filter_statement,
  list_elements = c("plot_lines", "plot_points", "plot_background_lines", "plot_braces")
)
}
\arguments{
\item{plot_dat}{A list of data.frames prepared by \code{\link[=prep_plot]{prep_plot()}}.}

\item{filter_statement}{Character string containing a logical expression for filtering specific rows in all data.frames in the list when possible. Write \code{dat$column_name} to specify the wanted columns.}

\item{list_elements}{Character vector with the names of the list objects you want to filter in. Defaults to \code{c("plot_lines", "plot_points", "plot_background_lines", "plot_braces")}.}
}
\value{
The plot_dat list with rows filtered as specified.
}
\description{
Filter specified rows from plot_dat objects.
}
\examples{
# tbd
}
