% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Extract or remove rows from prepared data.}
\usage{
filter_rows(
  plot_dat,
  column_name,
  subsetter,
  list_elements = c("plot_points", "plot_lines", "plot_braces"),
  remove = FALSE,
  remove_na = FALSE
)
}
\arguments{
\item{plot_dat}{List of data.frames, output of \link{prep_lineplot}.}

\item{column_name}{Column that the subsetter will be searched in.}

\item{subsetter}{Character string of the state you want to extract or remove.}

\item{list_elements}{Character vector with the names of the list objects you want to filter in. Defaults to \code{c("plot_lines", "plot_points", "plot_background_lines", "plot_braces")}.}

\item{remove}{Logical. If \code{TRUE}, the state defined in \code{state} will be removed. Defaults to \code{FALSE}.}

\item{remove_na}{Logical. If \code{TRUE}, rows with \code{NAs} in \code{column_name} will be removed.}
}
\value{
List of data.frames.
}
\description{
Extract or remove rows from prepared data.
}
\examples{
# tbd
}
