% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Extract or remove states from prepared data.}
\usage{
filter_rows(plot_data, state, remove = FALSE)
}
\arguments{
\item{plot_data}{List of data.frames, output of \link{prep_trend}.}

\item{state}{Character string of the state you want to extract or remove.}

\item{remove}{Logical. If true, the state defined in \code{state} will be removed. Defaults to \code{FALSE}.}
}
\value{
List of data.frames.
}
\description{
Extract or remove states from prepared data.
}
\examples{
#tbd
}
