% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub_plot_dat.R
\name{gsub_plot_dat}
\alias{gsub_plot_dat}
\title{\code{\link[=gsub]{gsub()}} for lists and data.frames.}
\usage{
gsub_plot_dat(
  plot_dat,
  search_for = "wholeGroup",
  replace_with = "Deutschland"
)
}
\arguments{
\item{plot_dat}{List or data.frame, mainly usage intention for plot_dat - objects from \code{\link[=prep_lineplot]{prep_lineplot()}}.}

\item{search_for}{Character string that should be replaced.}

\item{replace_with}{Character string that will be the replacement.}
}
\value{
Depending on input data.frame or list with the subbed character strings.
}
\description{
\code{\link[=gsub]{gsub()}} for lists and data.frames.
}
\examples{
plot_dat <- list(
  dat_1 = data.frame(
    col_1 = c("wholeGroup", "notwholeGroup"),
    col_2 = c("wholeGroup", NA),
    col_3 = c(1, 2)
  ),
  dat_2 = data.frame(
    col_1 = c("wholeGroup", "notwholeGroup"),
    col_2 = c("wholeGroup", NA),
    col_3 = c(TRUE, FALSE)
  )
)
gsub_plot_dat(plot_dat)
}
