% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_first_number}
\alias{insert_first_number}
\title{Insert a character string in front of the first number in a character string.}
\usage{
insert_first_number(char_string, insertion)
}
\arguments{
\item{char_string}{Character string that gets the insertion.}

\item{insertion}{Character to insert before the first number in \code{char_string}.}
}
\value{
A character string with an insertion in front of the first number.
}
\description{
Insert a character string in front of the first number in a character string.
}
\examples{
insert_first_number("test2013b", "\\\\.")
}
