% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineplot_3x2}
\alias{lineplot_3x2}
\title{BT24 default plot settings.}
\format{
\subsection{\code{lineplot_3x2}}{

A list with two elements.
\describe{
\item{n_cols}{Four columns of smaller plots are used for the final lineplot.}
\item{axis_x_label_centralize}{The x-axis labels are nudged into the center by 0.155 npc's.}
\item{split_plot}{}
\item{y_axis}{}
...
}
}
}
\usage{
lineplot_3x2
}
\description{
Default plot settings for plotting whole germany values over multiple competence areas.
}
\keyword{datasets}
