% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineplot_4x4}
\alias{lineplot_4x4}
\title{BT24 default plot settings.}
\format{
\subsection{\code{lineplot_4x4}}{

A list with two elements.
\describe{
\item{n_cols}{Four columns of smaller plots are used for the final lineplot.}
\item{axis_x_label_centralize}{The x-axis labels are nudged into the center by 0.155 npc's.}
\item{split_plot}{}
\item{y_axis}{}
...
}
}
}
\usage{
lineplot_4x4
}
\description{
Default plot settings for lineplots in the BT_22 for chapter 4.
}
\keyword{datasets}
