% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineplot_4x4_3groups}
\alias{lineplot_4x4_3groups}
\title{BT22 default plot settings.}
\format{
\subsection{\code{lineplot_4x4_3groups}}{

A list with two elements.
\describe{
\item{background_stripes_colour}{Colour of the background stripes.}
...
}
}
}
\usage{
lineplot_4x4_3groups
}
\description{
Default plot settings lineplots with three groups.
}
\keyword{datasets}
