% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lineplot_germany}
\alias{lineplot_germany}
\title{BT24 default plot settings.}
\format{
\subsection{\code{lineplot_germany}}{

A list with two elements.
\describe{
\item{point_shapes}{Point shapes with triangles as significant values.}
...
}
}
}
\usage{
lineplot_germany
}
\description{
Default plot settings for lineplots of whole germany in the BT_22 for chapter 4.
}
\keyword{datasets}
