% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{means}
\alias{means}
\title{BT24 example data.}
\format{
\subsection{\code{means}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
means
}
\description{
Example data that goes with the \code{anteile} data, but calculates its means.
}
\keyword{datasets}
