% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_trend_data.R
\name{merge_trend_data}
\alias{merge_trend_data}
\title{Wrapper for merging to trend dataframes by the year columns.}
\usage{
merge_trend_data(trend_data_1, trend_data_2, suffixes, ...)
}
\arguments{
\item{trend_data_1}{Trend data containing trend comparisons.}

\item{trend_data_2}{Trend data containing estimates for the trends.}

\item{suffixes}{Suffixes that should be added to duplicated columns.}
}
\value{
Merged data.frame stemming from the input data.frames.
}
\description{
Wrapper for merging to trend dataframes by the year columns.
}
\examples{
# tbd
}
