% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_trend_point.R
\name{merge_trend_point}
\alias{merge_trend_point}
\title{Merge point estimates onto the resepective timepoints.}
\usage{
merge_trend_point(trend_data, point_data)
}
\arguments{
\item{trend_data}{Data.frame with the trend comparisons.}

\item{point_data}{Data.frame with the point estimates for the start and end years.}
}
\value{
Data frame with the within significance estimates and the point estimates for start and end year.
}
\description{
Merge point estimates onto the resepective timepoints.
}
\examples{
#tbd
}
