% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_within_whole.R
\name{merge_within_whole}
\alias{merge_within_whole}
\title{Title}
\usage{
merge_within_whole(trend_data, groups, BLs)
}
\arguments{
\item{trend_data}{Trend data.}

\item{groups}{Groups.}

\item{BLs}{States.}
}
\value{
list of data frames.
}
\description{
Title
}
\examples{
#tbd
}
