% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_stand}
\alias{min_stand}
\title{BT21 minimum standards.}
\format{
\subsection{\code{min_stand}}{

A data frame with ... rows and ... columns:
\describe{
\item{kb}{Competence}
...
}
}
}
\source{
<"Q:/BT2022/BT/60_Bericht/_Probegrafiken/2023-01-26 Vorlagen Balken und Linien/BT2021_Abb3.9.xlsx">
}
\usage{
min_stand
}
\description{
BT21 data set that contains data for the mindest-standards.
}
\keyword{datasets}
