% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axis.R
\name{plot_axis}
\alias{plot_axis}
\title{Plot the x-Axis.}
\usage{
plot_axis(data_plot_points, y_range)
}
\arguments{
\item{data_plot_points}{Data frame with the point estimates for every year. Prepared by ...}

\item{y_range}{Numeric vector with two elements, indicating the min and max for the y axis.}
}
\value{
A blue stripe on top of ggplot2 plot which can be used as x-axis.
}
\description{
Plot the x-Axis.
}
\examples{
#tbd
}
