% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot an IQB barplot.}
\usage{
plot_bar(prep_dat)
}
\arguments{
\item{prep_dat}{Data for the barplot. Has to be in the right format, ideally prepared with \link{prep_barplot}.}
}
\value{
ggplot object
}
\description{
Plot an IQB barplot.
}
\details{
text describing parameter inputs in more detail.
\itemize{
\item{"prep_dat"}{Needs a specific data format.}
}
}
\examples{
#tbd
}
