% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_braces.R
\name{plot_braces}
\alias{plot_braces}
\title{Plot braces below plot.}
\usage{
plot_braces(data_trend_point)
}
\arguments{
\item{data_trend_point}{Prepared Trend data.}
}
\value{
ggplot2 object.
}
\description{
Plot braces below plot.
}
\examples{
##
}
