% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_column.R
\name{plot_column}
\alias{plot_column}
\title{Plot y-axis of a merged plot.}
\usage{
plot_column(
  vec,
  title = "y_value",
  ...,
  fontsize = 3,
  alignment = "left",
  nudge_x = -0.55
)
}
\arguments{
\item{vec}{Character vector of the strings that should be plotted as y-axis labels. E.g., states.}

\item{title}{Character string of the column title that will be ploted. Defaults to \code{"y_value"}}

\item{...}{Arguments for \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}

\item{fontsize}{Numeric for the fontsize of the text in the column. Defaults to \code{"3"}.}

\item{alignment}{Character string for the alignmen of the column text. Possible are \verb{"left", "center", "right"}. Defaults to \code{"left"}.}

\item{nudge_x}{Numeric for nudging the column text on the x-axis. Defaults to \code{-0.55}.}
}
\value{
ggplot2 object
}
\description{
Plot y-axis of a merged plot.
}
\examples{
# tbd
}
