% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_lineplot.R
\name{plot_single_lineplot}
\alias{plot_single_lineplot}
\title{Plot a single lineplot. Multiple of this lineplots are combined as tiles to the output of \code{\link[=plot_lineplot]{plot_lineplot()}}.}
\usage{
plot_single_lineplot(plot_dat)
}
\value{
ggplot2 Object.
}
\description{
Plot a single lineplot. Multiple of this lineplots are combined as tiles to the output of \code{\link[=plot_lineplot]{plot_lineplot()}}.
}
\examples{
# tbd
}
