% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Draw a table in ggplot2.}
\usage{
plot_table(data_prep)
}
\arguments{
\item{data_prep}{Data prepered with \code{prep_tableplot()}.}
}
\value{
ggplot2 object.
}
\description{
Draw a table in ggplot2.
}
\examples{
#tbd
}
