% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table_bar.R
\name{plot_table_bar}
\alias{plot_table_bar}
\title{Title}
\usage{
plot_table_bar(p_1, p_2)
}
\arguments{
\item{p_1}{Left plot. Currently a table_plot.}

\item{p_2}{Right plot. Currently a bar_plot.}
}
\value{
ggplot2 objet with both plots aligned as one.
}
\description{
Title
}
\examples{
#tbd
}
