% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table_continous.R
\name{plot_tablebar}
\alias{plot_tablebar}
\title{Plot a table and/or a barchart.}
\usage{
plot_tablebar(
  dat,
  bar_est = NULL,
  bar_label = NULL,
  bar_sig = NULL,
  bar_fill = NULL,
  bar_header = NULL,
  columns_headers = NULL,
  columns_table = NULL,
  columns_table_sig_bold = NULL,
  columns_table_sig_high = NULL,
  columns_table_se = NULL,
  y_axis = NULL,
  plot_settings = plotsettings_tablebarplot()
)
}
\arguments{
\item{dat}{Data prepared with \code{\link[=prep_trend]{prep_trend()}}.}

\item{bar_est}{Character string for the column that contains the values for the bar chart. If \code{NULL}, no bar chart will be plotted.}

\item{bar_label}{Character string for the column that should be used for bar labels on top of the bars. If \code{NULL}, no labels are printed. Defaults to \code{NULL}.}

\item{bar_sig}{Character string for the column that should be used for marking the bars as significant.}

\item{bar_fill}{Character string for the column that groups the bar filling colours into different groups.}

\item{bar_header}{Character string for the header of the bars.}

\item{columns_headers}{Character vector containing the column columns_headers.}

\item{columns_table}{List of character strings of the columns that should be plotted as table columns in the plot.}

\item{columns_table_sig_bold}{List of character strings of the columns that contain the significances for plotting significant values as bold.}

\item{columns_table_sig_high}{List of character strings of the columns that contain the significances for plotting significant values with a raised a.}

\item{columns_table_se}{List of character strings of the columns that contain standard errors, which will be plotted in brackets behind the column values.}

\item{y_axis}{Character string of the columnname used as y-axis.}

\item{plot_settings}{Named list constructed with \code{plotsettings_tablebarplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_tablebarplot()} for an overview.}
}
\value{
\link{ggplot2} object.
}
\description{
Plot a table and/or a barchart.
}
\examples{
#tbd
}
