% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tablebarplot.R
\name{plot_tablebarplot}
\alias{plot_tablebarplot}
\title{Plot a table and/or a barchart.}
\usage{
plot_tablebarplot(
  dat,
  y_axis = NULL,
  bar_est = NULL,
  bar_label = NULL,
  bar_sig = NULL,
  bar_label_sig = NULL,
  bar_fill = NULL,
  columns_table = NULL,
  columns_table_sig_bold = NULL,
  columns_table_sig_superscript = NULL,
  columns_table_se = NULL,
  headers = NULL,
  column_spanners = NULL,
  column_spanners_2 = NULL,
  columns_round = 0,
  plot_settings = plotsettings_tablebarplot()
)
}
\arguments{
\item{dat}{Data prepared with \code{\link[=prep_tablebarplot]{prep_tablebarplot()}}.}

\item{y_axis}{Character string of the columnname used as y-axis. Has to contain unique values.}

\item{bar_est}{Character string for the column that contains the values for the bar chart. If \code{NULL}, no bar chart will be plotted.}

\item{bar_label}{Character string for the column that should be used for bar labels on top of the bars. If \code{NULL}, no labels are printed. Defaults to \code{NULL}.}

\item{bar_sig}{Character string for the column that should be used for marking the bars as significant.}

\item{bar_label_sig}{Character string for the column that should be used for marking the bar labels as significant.}

\item{bar_fill}{Character string for the column that groups the bar filling colours into different groups.}

\item{columns_table}{List of character strings of the columns that should be plotted as table columns in the plot.}

\item{columns_table_sig_bold}{List of character strings of the columns that contain the significances for plotting significant values as bold.}

\item{columns_table_sig_superscript}{List of character strings of the columns that contain the significances for plotting significant values with a raised a.}

\item{columns_table_se}{List of character strings of the columns that contain standard errors, which will be plotted in brackets and rounded to \code{1}.}

\item{headers}{Character vector containing the headers of the ploted table columns, including the bar table.}

\item{column_spanners}{Named list. The name of each element will be the column header. The list element itself has to be a numeric vector indicating which columns the column spanner should span.}

\item{column_spanners_2}{Named list. A second dimension of column spanners. The name of each element will be the column header. The list element itself has to be a numeric vector indicating which columns the column spanner should span.}

\item{columns_round}{Numeric vector for rounding the column values. Insert \code{NULL} or \code{0} for no rounding/character columns.}

\item{plot_settings}{Named list constructed with \code{plotsettings_tablebarplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_tablebarplot()} for an overview.}
}
\value{
\code{ggplot2} object.
}
\description{
Plot a table and/or a barchart.
}
\examples{
# tbd
}
