% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_axis.R
\name{plot_y_axis}
\alias{plot_y_axis}
\title{Plot a y-axis plot. Can be merged with other plots for a customisable y-axis.}
\usage{
plot_y_axis(plot_dat, plot_lims, plot_settings = plotsettings_lineplot())
}
\arguments{
\item{plot_dat}{Input is a list prepared by \code{\link[=prep_plot]{prep_plot()}}.`}

\item{plot_lims}{List obtained by \code{\link[=calc_plot_lims]{calc_plot_lims()}}, containing different coordinate values.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
\link{ggplot2} plot that can be used as a y-axis.
}
\description{
Plot a y-axis plot. Can be merged with other plots for a customisable y-axis.
}
\examples{
# tbd
}
