% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsettings_barplot.R
\name{plotsettings_barplot}
\alias{plotsettings_barplot}
\title{Set parameters for the barplots.}
\usage{
plotsettings_barplot(
  background_stripes_colour = NULL,
  bar_fill_colour = NULL,
  bar_frame_linetype = NULL,
  bar_pattern_fill_colour = NULL,
  bar_pattern_type = NULL,
  bar_sig_type = NULL,
  default_list = NULL
)
}
\arguments{
\item{background_stripes_colour}{Named vector containing the two colours that should be used for the striped background.}

\item{bar_fill_colour}{Named vector with the bar filling colours. Names of the vector must be found in the column specified in \code{bar_fill}.#}

\item{bar_frame_linetype}{Named vector with the bar frame linetypes.}

\item{bar_pattern_fill_colour}{Named vector with the filling colours for the bar pattern. Names of the vector must be found in the column specified in \code{bar_pattern_fill}.}

\item{bar_pattern_type}{Named vector with the pattern types for the barpattern.}

\item{bar_sig_type}{Character string indicating whether levels of the grouping variable should be visualized by pattern fill ("pattern") or line type ("frame").}

\item{default_list}{Named list with predefined settings. Defaults to a list with all settings set to \code{0}.}
}
\value{
A named list with settings for the barplots.
}
\description{
Set parameters for the barplots.
}
\examples{
plotsettings_barplot()

}
