% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsettings_lineplot.R
\name{plotsettings_lineplot}
\alias{plotsettings_lineplot}
\title{Set parameters for the lineplots.}
\usage{
plotsettings_lineplot(
  axis_x_background_colour = NULL,
  axis_x_background_width_x = NULL,
  axis_x_background_width_y = NULL,
  axis_x_label_centralize = NULL,
  axis_x_label_nudge_y = NULL,
  axis_x_label_size = NULL,
  axis_y = NULL,
  axis_y_tick_distance = NULL,
  axis_y_lims = NULL,
  background_lines = NULL,
  background_line_colour = NULL,
  box_facet_linewidth = NULL,
  brace_label_gap_y = NULL,
  brace_label_nudge_x = NULL,
  brace_label_size = NULL,
  brace_label_nudge_y = NULL,
  brace_line_width = NULL,
  brace_span_y = NULL,
  equal_trend_line_length = NULL,
  line_type = NULL,
  line_width = NULL,
  margin_below_y_axis = NULL,
  margin_above_y_axis = NULL,
  margin_bottom = NULL,
  margin_left = NULL,
  margin_right = NULL,
  margin_top = NULL,
  n_cols = NULL,
  point_label_nudge = NULL,
  point_shapes = NULL,
  point_label_size = NULL,
  point_label_nudge_direction = NULL,
  point_label_nudge_x = NULL,
  point_label_nudge_y = NULL,
  point_size = NULL,
  split_plot = NULL,
  split_plot_gap_width = NULL,
  subgroup_colours = NULL,
  default_list = NULL
)
}
\arguments{
\item{axis_x_background_colour}{Colour value of the x-axis background.}

\item{axis_x_background_width_x, axis_x_background_width_y}{Numeric. The background space will be increased in x- or y-direction, if this parameter is increased.}

\item{axis_x_label_centralize}{Numeric. The x-axis labels will be nudged into the center by this amount, if the plot is a split lineplot.}

\item{axis_x_label_nudge_y}{Numeric for shifting the x-axis labels vertically. Increase to lower the x-axis labels.}

\item{axis_x_label_size}{Numeric for the font size of the x-axis labels.}

\item{axis_y}{Logical, indicating whether a y-axis should be plotted to the left of each row or not.}

\item{axis_y_tick_distance}{Numeric, for which distance should lie between tick marks. The first tick will start at the lower end of \code{axis_y_lims} or be calculated automatically. The following ticks will be added in distances defined by \code{axis_y_tick_distance}. Defaults to \code{20}.}

\item{axis_y_lims}{Numeric for the y-axis limits. Defaults to \code{NULL}, in which case the limits will be set automatically.}

\item{background_lines}{Logical, indicating whether the whole group trend should be plotted in the background. Defaults to \code{TRUE}.}

\item{background_line_colour}{Colour of the background line.}

\item{box_facet_linewidth}{Numeric for the linewidth of the box around chosen seperate_plot-vars (e.g., states).}

\item{brace_label_gap_y}{Numeric for the size of the vertical gap between brace labels.}

\item{brace_label_nudge_x}{Numeric. The brace labels will be shifted along the x-axis by this amount. Increase to shift the labels further to the right.}

\item{brace_label_size}{Numeric fontsize of the bracelabels.}

\item{brace_label_nudge_y}{Numeric. The brace labels will be shifted along the y-axis by this amount. Increase to let the labels start further below.}

\item{brace_line_width}{Numeric for the thickness of the brace.}

\item{brace_span_y}{Numeric for the width of the brace on the y-axis.}

\item{equal_trend_line_length}{Logical. If true, all trend lines will have the same length, even if the actual distance between the year is differently. E.g., the distance between the years of the trends 2010 - 2011 and 2011 - 2020 would be equal, even though the first trend only takes 1 year, and the second one 9 years. Please use with care, as it will distort the x-axis and results in misleading line lengths. Defaults to \code{FALSE}.}

\item{line_type}{Named vector containing the linetypes for the plots. The names must be either \code{"TRUE"} or \code{"FALSE"}, as the linetpyes represent significances. See \link[ggplot2:scale_linetype]{ggplot2::scale_linetype}.}

\item{line_width}{Numeric for the thicknes of the plotted lines.}

\item{margin_below_y_axis, margin_above_y_axis}{Numeric to scale the area between the y-axis and the braces/the x-axis-header.}

\item{margin_bottom, margin_left, margin_right, margin_top}{Numeric for the area around the plot. See \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{n_cols}{Numeric, indicating how many columns of smaller plots the final lineplot should have.}

\item{point_label_nudge}{Logical. If \code{TRUE}, the point labels will be nudged automatically by \link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel} to try and avoid them overlapping with lines and/or points. Might interfere with other point-label settings, so use at your own risk. Alternatively, you could try to use \code{point_label_nudge_direction} to get some controll over the point labels.}

\item{point_shapes}{Named numeric vector for the shape of the points. The vectornames must be either \code{"TRUE"} or \code{"FALSE"}, as the point-shapes refer to significances. See \link[ggplot2:scale_shape]{ggplot2::scale_shape}.}

\item{point_label_size}{Numeric for the fontsize of the pointlabels.}

\item{point_label_nudge_direction}{Named list with contents of either "+" or "-". The names have to be the factorlevels of the grouping_var. For "+" the point lables will be printed above the point, for "-" below. If \code{NULL} is provided, the labels will be printed below the points for the lowest group, and above the points for all others.}

\item{point_label_nudge_x, point_label_nudge_y}{Numeric for the amount the pointlabel is nudged in x- or y-direction.}

\item{point_size}{Numeric for the size of plotted points.}

\item{split_plot}{Logical, indicating whether the different trends should be split or not.}

\item{split_plot_gap_width}{Numeric for the width of the gap in a split plot in npc.}

\item{subgroup_colours}{Named vector with the colours for different grouping_var groups. The names of the vector have to be equivalent to the factorlevels of your grouping_var.}

\item{default_list}{Named list with predefined settings. Defaults to a list with all settings set to \code{0}.}
}
\value{
A named list with settings for the lineplots.
}
\description{
Set parameters for the lineplots.
}
\examples{
plotsettings_lineplot(n_cols = 2, axis_x_label_centralize = 0.1)
}
