% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_line.R
\name{prep_line}
\alias{prep_line}
\title{Title}
\usage{
prep_line(data, grouping_var, sig_niveau = 0.05)
}
\arguments{
\item{data}{Data for line plot.}

\item{grouping_var}{Variable to group by.}

\item{sig_niveau}{Significance niveau.}
}
\value{
Data frame in long format, which includes the relevant trend variables.
}
\description{
Title
}
\examples{
#tbd
}
