% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_lines.R
\name{prep_lines}
\alias{prep_lines}
\title{Merge point estimates onto the resepective timepoints.}
\usage{
prep_lines(prep_list)
}
\arguments{
\item{prep_list}{List stemming from \code{prep_lineplot()}.}
}
\value{
Data frame with the within significance estimates and the point estimates for start and end year.
}
\description{
Merge point estimates onto the resepective timepoints.
}
\examples{
#tbd
}
