% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_no_trend.R
\name{prep_no_trend}
\alias{prep_no_trend}
\title{Prepare data for plotting a barplot.}
\usage{
prep_no_trend(
  dat,
  columns,
  competence,
  grouping_var = "",
  group_var = "group",
  state_var = "TR_BUNDESLAND",
  competence_var = "kb",
  states = NULL,
  sig_niveau = 0.05,
  parameter = "mean"
)
}
\arguments{
\item{dat}{Input data.frame stemming from \code{eatRep}.}

\item{columns}{Character string containing the column in \code{data} that contains the column names for the table_plot.}

\item{competence}{Character string containing the competence that should be plotted.}

\item{grouping_var}{Character string containing the column name in \code{dat} that should be used to distinguish between subgroups.}

\item{group_var}{Character string containing the column name in \code{dat} that contains the different group memberships in one string. Defaults to \code{"group"}.}

\item{state_var}{Character string containing the column name in \code{dat} that should be used to distinguish between groups that should be plotted seperatly. Normally, this should be the states ("Bundesländer"). Therfore, defaults to \code{"TR_BUNDESLAND"}.}

\item{competence_var}{Character string containing the column name in \code{dat} that contains the different competences. Defaults to \code{"kb"}.}

\item{states}{Character vector of the states that should be plotted.}

\item{sig_niveau}{Numeric indicating the border, below which p-values will be considered significant. Defaults to \code{0.05}.}

\item{parameter}{Character string of the parameter value that should be reported. Defaults to \code{"mean"}.}
}
\value{
Returns a data.frame which can be used as input for plot_bar.
}
\description{
Prepare data for plotting a barplot.
}
\examples{
# tbd
}
