% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plot.R
\name{prep_plot}
\alias{prep_plot}
\title{Prepare trend data for plotting}
\usage{
prep_plot(
  dat,
  competence_var = "kb",
  competence = NULL,
  states = NULL,
  state_var = "TR_BUNDESLAND",
  group_var = "group",
  grouping_vars = NULL,
  grouping_vars_groups = NULL,
  sig_niveau = 0.05,
  plot_mean = FALSE,
  parameter = "mean"
)
}
\arguments{
\item{dat}{Input data.frame stemming from \code{eatRep}.}

\item{competence_var}{Character string containing the column name in \code{dat} that contains the different competences. Defaults to \code{"kb"}.}

\item{competence}{Character string containing the competence that should be plotted.}

\item{states}{Character vector of the states that should be plotted.}

\item{state_var}{Character string containing the column name in \code{dat} that should be used to distinguish between groups that should be plotted seperatly. Normally, this should be the states ("Bundesländer"). Therfore, defaults to \code{"TR_BUNDESLAND"}.}

\item{group_var}{Character string containing the column name in \code{dat} that contains the different group memberships in one string. Defaults to \code{"group"}.}

\item{grouping_vars}{Character vector containing maximal two column names in \code{dat} that should be used to distinguish between subgroups. If two columns are provided, they will be merged internally into one.}

\item{grouping_vars_groups}{Character vector containing the groups from \code{grouping_var} you want to plot. Defaults to \code{NULL}, in which case all groups are prepared.}

\item{sig_niveau}{Numeric indicating the border, below which p-values will be considered significant. Defaults to \code{0.05}.}

\item{plot_mean}{Logical value indicating whether the mean of the subgroups should be plotted as well.}

\item{parameter}{Character string of the parameter value that should be reported. Defaults to \code{"mean"}.}
}
\value{
\code{prep_plot()} returns a list containing four data.frames prepared for plotting with different \link{eatPlot} functions. This includes the data.frames:
\itemize{
\item \code{plot_points} for plotting with \code{\link[=plot_points]{plot_points()}}
\item \code{plot_lines} for plotting with \code{\link[=plot_lines]{plot_lines()}}
\item \code{plot_braces} for plotting with \code{\link[=plot_braces]{plot_braces()}}
\item \code{plot_background_lines} for plotting with \code{\link[=plot_background_lines]{plot_background_lines()}}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Performs different data transformations, to bring the input data.frame into the correct formats for different kind of plots.
}
\examples{
# tbd
}
