% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_points.R
\name{prep_points}
\alias{prep_points}
\title{Title}
\usage{
prep_points(data, competence, grouping_var, sig_niveau = 0.05)
}
\arguments{
\item{data}{Trend data.}

\item{competence}{Competence area.}

\item{grouping_var}{Variable that groups the cases.}

\item{sig_niveau}{When do p-values become significant?}
}
\value{
Data frame in long format, which includes the relevant point estimates.
}
\description{
Title
}
\examples{
# tbd
}
