% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tableplot.R
\name{prep_tableplot}
\alias{prep_tableplot}
\title{Prepare data to plot as table.}
\usage{
prep_tableplot(data, columns, competence, sig_niveau = 0.05)
}
\arguments{
\item{data}{Output from \code{eatRep}.}

\item{columns}{Column in data that contains the different columns that should be plottet.}

\item{competence}{Competence that should be plottet.}

\item{sig_niveau}{At which significance niveau do p-values become significant?}
}
\value{
Returns a \code{data.frame} which can be used as input for \code{plot_table}.
}
\description{
Prepare data to plot as table.
}
\examples{
# tbd
}
