% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_bundesland}
\alias{process_bundesland}
\title{Process the data in the Bundesland column.}
\usage{
process_bundesland(vec, linebreak = FALSE)
}
\arguments{
\item{linebreak}{Logical. If \code{TRUE}, a linebreak is added after the '-'. Defaults to \code{FALSE}.}

\item{eatPlot_dat}{Data.frame coming from \code{\link[=prep_lineplot]{prep_lineplot()}} or \code{\link[=prep_tablebarplot]{prep_tablebarplot()}}.}

\item{column}{Character string of the column containing the bundesl채nder. Defaults to \code{TR_BUNDESLAND}.}
}
\value{
The eatPlot_data data.frame with the bundesl채nder column processed.
}
\description{
This changes \code{ue} to \code{체} and adds a \code{-} between bundesl채nder consisting of two words.
}
\examples{
# tbd
}
