% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Title}
\usage{
save_plot(p, filename, width = 16, height = 25.2, format = "PDF", scaling = 1)
}
\arguments{
\item{p}{Plot that should be saved.}

\item{filename}{Filepath and -name for the new plot file.}

\item{width}{Width of the new plot file. Defaults to \code{16}.}

\item{height}{Height of the new plot file. Defaults to \code{25.2}}

\item{format}{Format of the new plot file. Currently, only PDF is needed.}

\item{scaling}{Proportional width and height changes, so the proportion between them is kept.}
}
\value{
A PDF file.
}
\description{
Title
}
\examples{
#tbd
}
