% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plot_coords.R
\name{set_plot_coords}
\alias{set_plot_coords}
\title{Set the x- and y-coordinates for a plot.}
\usage{
set_plot_coords(plot_data, plot_settings = plotsettings())
}
\arguments{
\item{plot_data}{Input is a list prepared by \code{\link[=prep_trend]{prep_trend()}}.`}

\item{plot_settings}{Named list constructed with \code{plot_settings()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plot_settings()} for an overview.}
}
\value{
\link{ggplot2} object.
}
\description{
Set the x- and y-coordinates for a plot.
}
\examples{
# tbd
}
