% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_plot.R
\name{settings_lineplot}
\alias{settings_lineplot}
\title{Set theme, colours, pointshapes and linetypes for lineplot.}
\usage{
settings_lineplot(data_trend_point)
}
\arguments{
\item{time_points}{Data in Long format}
}
\value{
ggplot2 plot settings
}
\description{
Set theme, colours, pointshapes and linetypes for lineplot.
}
\examples{
#tbd
}
