% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_schemes.R
\name{sig_linetypes}
\alias{sig_linetypes}
\title{Wrapper around \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} with two predefined linetypes.}
\usage{
sig_linetypes()
}
\arguments{
\item{...}{Arguments for \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}.}
}
\value{
ggplot2
}
\description{
Wrapper around \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}} with two predefined linetypes.
}
\details{
Significant lines are solid, not-significant lines are dashed, according to the respective p-values.
}
\examples{
#tbd
}
