% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_schemes.R
\name{sig_pointshapes}
\alias{sig_pointshapes}
\title{Wrapper around \link[ggplot2]{scale_manual} with two predefined shapes.}
\usage{
sig_pointshapes(...)
}
\arguments{
\item{...}{Arguments for \link[ggplot2]{scale_manual}.}
}
\value{
ggplot2
}
\description{
Wrapper around \link[ggplot2]{scale_manual} with two predefined shapes.
}
\details{
Turns points int rectangles if the according p-value is significant.
}
\examples{
#tbd
}
