% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_column_width.R
\name{standardize_column_width}
\alias{standardize_column_width}
\title{Calculate the column widths for plots that should be later combined with \link{combine_plots}.}
\usage{
standardize_column_width(column_widths, plot_ranges)
}
\arguments{
\item{column_widths}{List of numeric vectors with proportional column widths regarding the final, combined plot. In the end, all proportions have to sum up to 1. However, if you want to set the width of barplots automatically, you can provide an \code{NA} for the column containing the barplot. In this case, it's width will be calculated automatically, so the proportions stay the same.}

\item{plot_ranges}{Numeric vector containing the ranges of the x-axis for alle barplots.}
}
\value{
Returns a list with numeric vectors containing the relative column widths that have to be set in the single plots.
}
\description{
Calculate the column widths for plots that should be later combined with \link{combine_plots}.
}
\examples{

standardize_column_width(column_widths = list(p1 = c(0.1, 0.2),
                                              p2 = c(0.5, 0.2)
                                              ),
                         plot_ranges = c(10, 30))

## NAs will be interpreted as barplots, in wich case the width of the barplots will be
## calclulated automatically, so the x-axes are on the same scale.
standardize_column_width(column_widths = list(p1 = c(0.1, NA),
                                              p2 = c(0.5, NA)
                                              ),
                         plot_ranges = c(10, 30))
}
