% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_pattern_spacing.R
\name{standardize_pattern_spacing}
\alias{standardize_pattern_spacing}
\title{Calculate the pattern spacing, so it is equal within a plot. Currently works only for two combined plots!}
\usage{
standardize_pattern_spacing(plot_list, pattern_spacing = 0.015)
}
\arguments{
\item{plot_list}{List of plots that will be combined by \code{\link[=combine_plots]{combine_plots()}}.}

\item{pattern_spacing}{Pattern spacing that will be taken as default for a whole plot. Defaults to \code{0.015}.}
}
\value{
Returns a numeric vector containing the pattern spacings for the plots.
}
\description{
This function takes a pattern spacing, that is for example used for single plots, and calculates the relative pattern_spacings needed for combined plots.
}
\examples{
# 1) Build the plots p1, p2.
# 2) standardize_pattern_spacing(list(p1, p2))
# 3) Build p1 and p2 new, but with the pattern spacings calculated in step 2.
}
