% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_line}
\alias{theme_line}
\title{Theme for lineplot.}
\usage{
theme_line(plot_settings = plotsettings_lineplot())
}
\arguments{
\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
ggplot2 theme
}
\description{
Theme for lineplot.
}
\examples{
# tbd
}
