% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_table}
\alias{theme_table}
\title{Theme for table plotted with ggplot2.}
\usage{
theme_table(plot_settings)
}
\arguments{
\item{plot_settings}{Named list constructed with \code{plotsettings_tablebarplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_tablebarplot()} for an overview.}
}
\value{
ggplot2 theme.
}
\description{
Theme for table plotted with ggplot2.
}
\examples{
# tbd
}
