% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_table_bar}
\alias{theme_table_bar}
\title{Theme for barplot that will be combined with a table.}
\usage{
theme_table_bar()
}
\value{
ggplot2 theme.
}
\description{
Theme for barplot that will be combined with a table.
}
