% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_table_col}
\alias{theme_table_col}
\title{Theme for a table column.}
\usage{
theme_table_col()
}
\value{
\code{ggplot2} theme, that can be used for a plotted table column.
}
\description{
Theme for a table column.
}
\examples{
# tbd
}
