% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trend_2}
\alias{trend_2}
\title{BT24 example data.}
\format{
\subsection{\code{trend_2}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
trend_2
}
\description{
Default plot settings for mean-barplot without trend.
}
\keyword{datasets}
