% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trend_books}
\alias{trend_books}
\title{BT21 Books trend}
\format{
\subsection{\code{trend_books}}{

A data frame with 2904 rows and 32 columns:
\describe{
\item{kb}{Competence}
...
}
}
}
\source{
<"Q:/BT2021/BT/60_Bericht/06_Soziale_Disparitäten/01_Rechnungen/KAS/Abb65Buecher_mitTrend.csv">
}
\usage{
trend_books
}
\description{
BT21 data set that contains some results published in the social disparities chapter.
}
\keyword{datasets}
