% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trend_gender}
\alias{trend_gender}
\title{BT24 example data.}
\format{
\subsection{\code{trend_gender}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
trend_gender
}
\description{
Example data with grouping var with multiple levels.
}
\keyword{datasets}
