% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trend_mw}
\alias{trend_mw}
\title{BT24 example data.}
\format{
\subsection{\code{trend_mw}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
trend_mw
}
\description{
Example data with male/female subgroup_var.
}
\keyword{datasets}
