% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trend_zwh}
\alias{trend_zwh}
\title{BT24 example data.}
\format{
\subsection{\code{trend_zwh}}{

A list with four elements.
\describe{
...
}
}
}
\usage{
trend_zwh
}
\description{
Example data with zuwanderungshintergrund as subgroup_var.
}
\keyword{datasets}
