% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{build_columns}
\alias{build_columns}
\title{Build columns from geom_text}
\usage{
build_columns(data_prep)
}
\arguments{
\item{data_prep}{Data prepared with \code{prep_tableplot()}.}
}
\value{
ggplot2 object
}
\description{
Build columns from geom_text
}
\examples{
#tbd
}
