% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_MinStand.R
\name{calc_MinStand}
\alias{calc_MinStand}
\title{Calculate the significance descriptions for the 'Mindeststandards'}
\usage{
calc_MinStand(dat_bar, sig_column, est_column)
}
\arguments{
\item{dat_bar}{Data prepared by \code{prep_plot()}.}

\item{sig_column}{Column that contains the significances.}

\item{est_column}{Column that contains the estimates.}
}
\value{
Dataframe with added \code{sig_minstand} column.
}
\description{
Calculate the significance descriptions for the 'Mindeststandards'
}
\examples{
#tbd
}
