% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Fill in and rename necessary columns}
\usage{
clean_data(
  dat,
  all_states,
  states = NULL,
  sub_groups = NULL,
  competence = NULL,
  parameter = "mean"
)
}
\arguments{
\item{dat}{Input data.frame stemming from \code{eatRep}.}

\item{all_states}{Character vector of the different groups that should be plotted seperatly. Normally these should be the states (Bundesländer).}

\item{states}{Character vector of the states that should be plotted.}

\item{sub_groups}{Character vector of the different sub_groups.}

\item{competence}{Character string containing the competence that should be plotted.}

\item{parameter}{Character string of the parameter value that should be reported. Defaults to \code{"mean"}.}
}
\value{
\code{clean_data()} returns a \code{data.frame} with renamend columns and filled in \code{NAs}.
}
\description{
\code{clean_data()} performs some simple data wrangling to prepare the input \code{data.frame} for the more plot specific data transformations.
}
\details{
The following operations are performed on \code{dat}:
\itemize{
\item For consistency, points in column names are subbed with underscores.
\item Only rows containing \code{mean} in \code{dat$parameter} are returend.
\item The group \code{wholeGroup} currently contains \code{NAs} in the columns \code{dat[ , state_var]} and \code{dat$grouping_var}. These are subbed with \code{wholeGroup} or \code{noGroup} respectively.
\item The column \code{dat[ , state_var]} is filled up by extracting the first state found in \code{dat$groups} for the respective row.
}
}
\examples{
# tbd
}
