% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_plots.R
\name{combine_plots}
\alias{combine_plots}
\title{Combine different ggplots}
\usage{
combine_plots(plot_list)
}
\arguments{
\item{plot_list}{List of the plots that should be combined.}
}
\value{
Patchwork plot.
}
\description{
Combine different ggplots
}
\examples{
#tbd
}
