% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_points.R
\name{connect_points}
\alias{connect_points}
\title{Connect Points}
\usage{
connect_points(my_data, grouping_var)
}
\arguments{
\item{my_data}{Data}

\item{grouping_var}{Variable for grouping}
}
\value{
ggplot object
}
\description{
Connect Points
}
