% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_grouping_var.R
\name{construct_grouping_var}
\alias{construct_grouping_var}
\title{Merge two grouping var columns into one.}
\usage{
construct_grouping_var(dat, grouping_vars, group_var = "group")
}
\arguments{
\item{dat}{Data frame with a \code{group} column.}

\item{grouping_vars}{Character vector containing maximal two column names in \code{dat} that should be used to distinguish between subgroups. If two columns are provided, they will be merged internally into one.}

\item{group_var}{Name of the \code{group} column in your data. Defaults to \code{group}.}
}
\value{
Data frame with a new \code{grouping_var} column, constructed from the provided \code{grouping_vars} and the \code{group_var} column.
}
\description{
Merge two grouping var columns into one.
}
\examples{
# tbd
}
