% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_label.R
\name{construct_label}
\alias{construct_label}
\title{Build labels with significances represented in bold or as raised 'a', and, if wanted, standard errors in brackets. Main usage for plotting tables and brace labels.}
\usage{
construct_label(
  dat,
  label_est = NULL,
  label_se = NULL,
  label_sig_bold = NULL,
  label_sig_high = NULL,
  round_est = 0,
  round_se = 1
)
}
\arguments{
\item{dat}{Dataframe with the columns that should be merged into labels.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'.}

\item{round_est}{Rounding of label_est.}

\item{round_se}{Rounding of label_se.}
}
\value{
A character vector with the constructed labels.
}
\description{
Build labels with significances represented in bold or as raised 'a', and, if wanted, standard errors in brackets. Main usage for plotting tables and brace labels.
}
\examples{
#tbd
}
