% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_label.R
\name{construct_label}
\alias{construct_label}
\title{Construct labels}
\usage{
construct_label(
  dat,
  column_est = NULL,
  column_se = NULL,
  column_sig_bold = NULL,
  column_sig_superscript = NULL,
  sig_superscript_letter = NULL,
  round_est = 0,
  round_se = 1,
  plot_settings = plotsettings_tablebarplot()
)
}
\arguments{
\item{dat}{Data frame with the columns that should be merged into labels.}

\item{column_est}{Character string for the column with the estimates.}

\item{column_se}{Character string for the column with the standard errors.}

\item{column_sig_bold}{Character string for the column with the significant estimates that should determine the bold printing.}

\item{column_sig_superscript}{Character string for the column with the significant estimates that should determine the superscript.}

\item{sig_superscript_letter}{Character string for the letter that should be used for the superscript.}

\item{round_est}{Rounding of brace_label_est.}

\item{round_se}{Rounding of brace_label_se.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
The data frame with an added column for the constructed label.
}
\description{
This function creates a new column or label, merging estimates and standard errors with significant estimates represented in bold or as superscript (via \code{brace_label_est}), and standard errors in brackets, if needed (via \code{brace_label_se}). NAs are converted to empty strings. Main usage is for plotting tables and brace labels.
}
\examples{
# example data frame
dat <- data.frame(
  names = c("Berlin", "Hamburg", "Hessen", "Niedersachsen", "Saarland"),
  estimate = c(400, 650, 380, 500, 600),
  se = c(0.1, 0.45, 1, 0.27, 0.9),
  p_estimate = c(FALSE, FALSE, TRUE, TRUE, FALSE)
)
}
