% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eatPlot-package.R
\docType{package}
\name{eatPlot-package}
\alias{eatPlot}
\alias{eatPlot-package}
\title{eatPlot: Plot eatRep output}
\description{
Functions for plotting eatModel output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nickhaf.github.io/eatPlot/}
}

}
\author{
\strong{Maintainer}: Nicklas Hafiz \email{hafiznij@hu-berlin.de}

Other contributors:
\itemize{
  \item Philipp Franikowski [contributor]
}

}
\keyword{internal}
