% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{Extract or remove rows from prepared data.}
\usage{
filter_rows(plot_data, column_name, subsetter, remove = FALSE)
}
\arguments{
\item{plot_data}{List of data.frames, output of \link{prep_plot}.}

\item{column_name}{Column that the subsetter will be searched in.}

\item{subsetter}{Character string of the state you want to extract or remove.}

\item{remove}{Logical. If true, the state defined in \code{state} will be removed. Defaults to \code{FALSE}.}
}
\value{
List of data.frames.
}
\description{
Extract or remove rows from prepared data.
}
\examples{
#tbd
}
