% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_background_lines.R
\name{plot_background_lines}
\alias{plot_background_lines}
\title{Plot background lines.}
\usage{
plot_background_lines(dat_total, plot_settings = plotsettings_lineplot())
}
\arguments{
\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}

\item{dat}{Data.}
}
\value{
ggplot2 object.
}
\description{
Plot background lines.
}
\examples{
# tbd
}
