% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_background_lines.R
\name{plot_background_lines}
\alias{plot_background_lines}
\title{Plot background lines.}
\usage{
plot_background_lines(data_plot_background_lines)
}
\arguments{
\item{data_plot_background_lines}{Data.}
}
\value{
ggplot2 object.
}
\description{
Plot background lines.
}
\examples{
#tbd
}
