% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot an IQB barplot.}
\usage{
plot_bar(
  no_trend_list,
  x_value = "est_wholeGroup",
  y_value = "state_var",
  bar_pattern = "sig_wholeGroup",
  bar_fill = "fill_wholeGroup",
  bar_pattern_fill = "grouping_var",
  bar_pattern_setting = sig_pattern,
  bar_fill_setting = adj_fill,
  bar_pattern_fill_setting = adj_pattern_fill,
  sig_pattern = "pattern"
)
}
\arguments{
\item{no_trend_list}{Input is a list prepared by \code{\link[=prep_no_trend]{prep_no_trend()}}. You can also use the according data.frame named \code{plot_bar} from this list.}

\item{x_value}{Character string of the column name containing the estimates that should be plotted on the x-axis. Defaults to \code{"est_wholeGroup"}, which are the estimates for the comparison of a state against the wholeGroup (Germany).}

\item{y_value}{Character string of the column name containing the labels that should be plotted on the y-axis. Defaults to \code{"state_var"}, so the states (Bundesländer) are depicted on the y-axis.}

\item{bar_pattern}{Character string of the column containing the grouping for the pattern on the bar. Defaults to \code{sig_wholeGroup}, so the significances of the state vs. wholeGroup (Germany) comparison are represented in the pattern of the bars.}

\item{bar_fill}{Character string of the column containing the grouping for the filling of the bar. Defaults to \code{fill_wholeGroup}, so the significances of the state vs. wholeGroup (Germany) comparison, as well as the groups found in "data$grouping_var" are represented in the filling colours of the bars.}

\item{bar_pattern_fill}{Character string of the column containing the grouping for the filling of the pattern on the bar. Defaults to \code{fill_wholeGroup}, so the groups found in "data$grouping_var" are represented in the colours of the bar pattern.}

\item{bar_pattern_setting}{Named vector with the pattern types. Names of the vector must be found in the column specified in \code{bar_pattern}. Defaults to ...}

\item{bar_fill_setting}{Named vector with the bar filling colours. Names of the vector must be found in the column specified in \code{bar_fill}. Defaults to ...}

\item{bar_pattern_fill_setting}{Named vector with the filling colours for the bar pattern. Names of the vector must be found in the column specified in \code{bar_pattern_fill}. Defaults to ...}

\item{sig_pattern}{Character string indicating whether groups should be rendered by pattern fill ("pattern") or line type ("frame"). Defaults to "pattern".}
}
\value{
Returns a \link{ggplot2} barplot.
}
\description{
Plot an IQB barplot.
}
\examples{
#tbd
}
