% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot an IQB barplot.}
\usage{
plot_bar(
  no_trend_list,
  x_value = "est_wholeGroup",
  y_value = "state_var",
  bar_label = "est_wholeGroup",
  bar_sig = "sig_wholeGroup",
  bar_fill = "fill_wholeGroup",
  bar_pattern_fill = "grouping_var",
  plot_settings = plotsettings_barplot()
)
}
\arguments{
\item{no_trend_list}{Input is a list prepared by \code{\link[=prep_no_trend]{prep_no_trend()}}. You can also use the according data.frame named \code{plot_bar} from this list.}

\item{x_value}{Character string of the column name containing the estimates that should be plotted on the x-axis. Defaults to \code{"est_wholeGroup"}, which are the estimates for the comparison of a state against the wholeGroup (Germany).}

\item{y_value}{Character string of the column name containing the labels that should be plotted on the y-axis. Defaults to \code{"state_var"}, so the states (Bundesländer) are depicted on the y-axis.}

\item{bar_label}{Character string of the column that should be used as label at the bar end. Defaults to \code{est_wholeGroup}. If \code{NULL}, no label will be plotted.}

\item{bar_sig}{Character string of the column containing the grouping for the pattern or the frame of the bar. Defaults to \code{sig_wholeGroup}, so the significances of the state vs. wholeGroup (Germany) comparison are represented in the pattern or the frame of the bars.}

\item{bar_fill}{Character string of the column containing the grouping for the filling of the bar. Defaults to \code{fill_wholeGroup}, so the significances of the state vs. wholeGroup (Germany) comparison, as well as the groups found in "data$grouping_var" are represented in the filling colours of the bars.}

\item{bar_pattern_fill}{Character string of the column containing the grouping for the filling of the pattern on the bar. Defaults to \code{grouping_var}, so the groups found in \code{data$grouping_var} are represented in the colours of the bar pattern.}

\item{plot_settings}{Named list constructed with \code{plotsettings_barplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_barplot()} for an overview.}
}
\value{
Returns a \link{ggplot2} barplot.
}
\description{
Plot an IQB barplot.
}
\examples{
#tbd
}
