% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_braces.R
\name{plot_braces}
\alias{plot_braces}
\title{Plot braces below plot.}
\usage{
plot_braces(
  dat,
  plot_lims,
  label_est = NULL,
  label_se = NULL,
  label_sig_high = NULL,
  label_sig_bold = NULL,
  plot_settings = plotsettings_lineplot()
)
}
\arguments{
\item{dat}{Prepared Trend data.}

\item{plot_lims}{List obtained by \code{\link[=calc_plot_lims]{calc_plot_lims()}}, containing different coordinate values.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'. Normally, should be the comparison of the trend vs. the trend in whole Germany, which can be found in the trendDiff_cross parameter. Defaults to \code{NULL}, as this parameter is not always provided.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold. Defaults to \code{"sig_Trend_noComp"}.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
ggplot2 object.
}
\description{
Plot braces below plot.
}
\examples{
##
}
