% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_braces.R
\name{plot_braces}
\alias{plot_braces}
\title{Plot braces below plot.}
\usage{
plot_braces(
  dat,
  y_range,
  label_est = NULL,
  label_se = NULL,
  label_sig_high = NULL,
  label_sig_bold = NULL,
  plot_settings = plotsettings_lineplot()
)
}
\arguments{
\item{dat}{Prepared Trend data.}

\item{y_range}{Numeric vector with two elements, indicating the min and max for the y axis.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'. Defaults to \code{sig_Trend_Comp_crossDiff_wholeGroupSameGroup}.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
ggplot2 object.
}
\description{
Plot braces below plot.
}
\examples{
##
}
