% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_braces.R
\name{plot_braces}
\alias{plot_braces}
\title{Plot braces below plot.}
\usage{
plot_braces(
  data_plot_braces,
  y_range,
  BL,
  label_est,
  label_se,
  label_sig_high,
  label_sig_bold
)
}
\arguments{
\item{data_plot_braces}{Prepared Trend data.}

\item{BL}{Bundesland, for which the labels should be drawn.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold.}
}
\value{
ggplot2 object.
}
\description{
Plot braces below plot.
}
\examples{
##
}
