% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lineplot.R
\name{plot_lineplot}
\alias{plot_lineplot}
\title{Title}
\usage{
plot_lineplot(
  plot_data,
  seperate_plot_var = "state_var",
  point_values = "est_point",
  point_sig = "sig_point",
  line_values = c("est_point_start", "est_point_end"),
  line_sig = "sig_trend_comp_within",
  label_est = "est_trend_no_comp",
  label_se = "se_trend_no_comp",
  label_sig_high = "sig_trend_comp_whole",
  label_sig_bold = "sig_trend_no_comp",
  plot_settings = plotsettings()
)
}
\arguments{
\item{plot_data}{Input is a list prepared by \code{\link[=prep_trend]{prep_trend()}}.`}

\item{seperate_plot_var}{Character string of the column containing the tiles. For every unique value, a new tile will be plotted. Defaults to \code{state_var}.}

\item{point_values}{Character string of the column name in \code{plot_data[["plot_points"]]} containing the y-values for the plotted points. Defaults to \code{est_point}.}

\item{point_sig}{Character string of the column name containing significance values for \code{point_values}. Defaults to \code{"sig_point"}.}

\item{line_values}{Character vector with two elements. Column names in \code{plot_data[["plot_lines"]]} containing the y-values for the plotted lines. Defaults to \code{c("est_point_start", "est_point_end")}.}

\item{line_sig}{Character string of the column name containing significance values for \code{line_values}. Defaults to \code{"sig_trend_comp_within"}.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold.}

\item{plot_settings}{Named list constructed with \code{plot_settings()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plot_settings()} for an overview.}
}
\value{
\link{ggplot2} object.
}
\description{
Title
}
\examples{
# tbd
}
