% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Adds a layer with trend lines to a ggplot.}
\usage{
plot_lines(data_lines)
}
\arguments{
\item{data_lines}{}
}
\value{
ggplot2 object
}
\description{
Adds a layer with trend lines to a ggplot.
}
\details{
Wrapper for \link[ggplot2]{geom_segment}.
}
\examples{
#tbd
}
