% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lines.R
\name{plot_lines}
\alias{plot_lines}
\title{Adds a layer with trend lines to a ggplot.}
\usage{
plot_lines(data_plot_lines)
}
\arguments{
\item{data_plot_lines}{Output \code{data.frame} from \code{\link[=prep_trend_point]{prep_trend_point()}}.}
}
\value{
ggplot2 object
}
\description{
Adds a layer with trend lines to a ggplot.
}
\details{
Wrapper for \code{\link[ggplot2:geom_segment]{ggplot2::geom_segment()}}.
}
\examples{
#tbd
}
