% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Plot Points.}
\usage{
plot_points(plot_dat)
}
\arguments{
\item{data_plot_points}{Data frame with the point estimates for every year. Prepared by ...}
}
\value{
ggplot2 object
}
\description{
Plot Points.
}
\examples{
# tbd
}
