% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_points.R
\name{plot_points}
\alias{plot_points}
\title{Plot Points.}
\usage{
plot_points(grouping_var)
}
\arguments{
\item{grouping_var}{Indicating which groups are getting distinct colours.}
}
\value{
ggplot2 object
}
\description{
Plot Points.
}
\examples{
#tbd
}
