% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_single_lineplot.R
\name{plot_single_lineplot}
\alias{plot_single_lineplot}
\title{Plot a single lineplot. Multiple of this lineplots are combined as tiles to the output of \code{\link[=plot_lineplot]{plot_lineplot()}}.}
\usage{
plot_single_lineplot(
  plot_data,
  y_range = NULL,
  point_values = "est_point",
  point_sig = "sig_point",
  line_values = c("est_point_start", "est_point_end"),
  line_sig = "sig_trend_comp_within",
  label_est = "est_trend_no_comp",
  label_se = "se_trend_no_comp",
  label_sig_high = "sig_trend_comp_whole",
  label_sig_bold = "sig_trend_no_comp",
  background_lines = TRUE,
  plot_settings = plotsettings()
)
}
\arguments{
\item{plot_data}{Input is a list prepared by \code{\link[=prep_trend]{prep_trend()}}.`}

\item{y_range}{Numeric vector with two elements, indicating the min and max for the y axis.}

\item{point_values}{Character string of the column name in \code{plot_data[["plot_points"]]} containing the y-values for the plotted points. Defaults to \code{est_point}.}

\item{point_sig}{Character string of the column name containing significance values for \code{point_values}. Defaults to \code{"sig_point"}.}

\item{line_values}{Character vector with two elements. Column names in \code{plot_data[["plot_lines"]]} containing the y-values for the plotted lines. Defaults to \code{c("est_point_start", "est_point_end")}. If set to \code{NULL}, no lines will be plotted.}

\item{line_sig}{Character string of the column name containing significance values for \code{line_values}. Defaults to \code{"sig_trend_comp_within"}.}

\item{label_est}{Character string of the column name containing the brace labels.}

\item{label_se}{Character string of the column name containing the standard errors for \code{label_est}. Will be put in bracktes behind \code{label_est}.}

\item{label_sig_high}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked by a raised 'a'.}

\item{label_sig_bold}{Character string of the column name containing significance values for \code{label_est}. Significant values will be marked as bold.}

\item{background_lines}{Logical, indicating whether the whole group trend should be plotted in the background.}

\item{plot_settings}{Named list constructed with \code{plot_settings()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plot_settings()} for an overview.}
}
\value{
ggplot2 Object.
}
\description{
Plot a single lineplot. Multiple of this lineplots are combined as tiles to the output of \code{\link[=plot_lineplot]{plot_lineplot()}}.
}
\examples{
# tbd
}
