% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x_axis.R
\name{plot_x_axis}
\alias{plot_x_axis}
\title{Plot the x-Axis.}
\usage{
plot_x_axis(plot_dat)
}
\value{
A blue stripe on top of ggplot2 plot which can be used as x-axis.
}
\description{
Plot the x-Axis.
}
\examples{
# tbd
}
