% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_x_axis.R
\name{plot_x_axis}
\alias{plot_x_axis}
\title{Plot the x-Axis.}
\usage{
plot_x_axis(
  data_plot_points,
  plot_lims,
  plot_settings = plotsettings_lineplot()
)
}
\arguments{
\item{data_plot_points}{Data frame with the point estimates for every year. Prepared by ...}

\item{plot_lims}{List obtained by \code{\link[=calc_plot_lims]{calc_plot_lims()}}, containing different coordinate values.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
A blue stripe on top of ggplot2 plot which can be used as x-axis.
}
\description{
Plot the x-Axis.
}
\examples{
# tbd
}
