% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_axis.R
\name{plot_y_axis}
\alias{plot_y_axis}
\title{Plot y-axis of a merged plot.}
\usage{
plot_y_axis(vec)
}
\arguments{
\item{vec}{Character vector of the strings that should be plotted as y-axis labels. E.g., states.}
}
\value{
ggplot2 object
}
\description{
Plot y-axis of a merged plot.
}
\examples{
# tbd
}
