% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_axis.R
\name{plot_y_axis}
\alias{plot_y_axis}
\title{Plot a y-axis plot. Can be merged with other plots for a customisable y-axis.}
\usage{
plot_y_axis(plot_dat, plot_settings = plotsettings_lineplot())
}
\arguments{
\item{plot_dat}{Input is a list prepared by \code{prep_lineplot()}.}
}
\value{
\link{ggplot2} plot that can be used as a y-axis.
}
\description{
Plot a y-axis plot. Can be merged with other plots for a customisable y-axis.
}
\examples{
# tbd
}
