% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsettings.R
\name{plotsettings}
\alias{plotsettings}
\title{Set parameters for the lineplots.}
\usage{
plotsettings(
  n_cols = NULL,
  nudge_x_axis = NULL,
  split_plot = NULL,
  y_axis = NULL,
  default_list = NULL
)
}
\arguments{
\item{n_cols}{Numeric, indicating how many columns of smaller plots the final lineplot should have. Defaults to \code{4}.}

\item{nudge_x_axis}{Numeric. The x-axis labels will be nudged into the center by this amount, if the plot is a split lineplot. Defaults to \code{0.4}.}

\item{split_plot}{Logical, indicating whether the different trends should be split or not.}

\item{y_axis}{Logical, indicating whether a y-axis should be plotted to the left of each row or not. Defaults to \code{FALSE}.}

\item{default_list}{Named list with predefined settings. Defaults to a list with all settings set to \code{0}.}
}
\value{
A named list with settings for the lineplots.
}
\description{
Set parameters for the lineplots.
}
\examples{
plotsettings(n_cols = 2, nudge_x_axis = 0.1)
}
