% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsettings.R
\name{plotsettings}
\alias{plotsettings}
\title{Set parameters for the lineplots.}
\usage{
plotsettings(
  axis_x_background_colour = NULL,
  axis_x_background_width_x = NULL,
  axis_x_background_width_y = NULL,
  axis_x_label_centralize = NULL,
  axis_x_label_nudge_y = NULL,
  axis_x_label_size = NULL,
  brace_label_gap_y = NULL,
  brace_label_nudge_x = NULL,
  brace_label_size = NULL,
  brace_label_nudge_y = NULL,
  brace_line_width = NULL,
  brace_span_y = NULL,
  grouping_colours = NULL,
  line_width = NULL,
  margin_bottom = NULL,
  margin_left = NULL,
  margin_right = NULL,
  margin_top = NULL,
  n_cols = NULL,
  point_label_nudge = NULL,
  point_label_size = NULL,
  point_label_nudge_direction = NULL,
  point_label_nudge_x = NULL,
  point_label_nudge_y = NULL,
  point_size = NULL,
  split_plot = NULL,
  split_plot_gap_width = NULL,
  y_axis = NULL,
  default_list = NULL
)
}
\arguments{
\item{axis_x_background_colour}{Colour value of the x-axis background.}

\item{axis_x_background_width_x, axis_x_background_width_y}{Numeric. The background space will be increased in x- or y-direction, if this parameter is increased.}

\item{axis_x_label_centralize}{Numeric. The x-axis labels will be nudged into the center by this amount, if the plot is a split lineplot.}

\item{axis_x_label_nudge_y}{Numeric for shifting the x-axis labels vertically. Increase to lower the x-axis labels.}

\item{axis_x_label_size}{Numeric for the font size of the x-axis labels.}

\item{brace_label_gap_y}{Numeric for the size of the vertical gap between brace labels.}

\item{brace_label_nudge_x}{Numeric. The brace labels will be shifted along the x-axis by this amount. Increase to shift the labels further to the right.}

\item{brace_label_size}{Numeric fontsize of the bracelabels.}

\item{brace_label_nudge_y}{Numeric. The brace labels will be shifted along the y-axis by this amount. Increase to let the labels start further below.}

\item{brace_line_width}{Numeric for the thickness of the brace.}

\item{brace_span_y}{Numeric for the width of the brace on the y-axis.}

\item{grouping_colours}{Named vector with the colours for different grouping_var groups. The names of the vector have to be equivalent to the factorlevels of your grouping_var.}

\item{line_width}{Numeric for the thicknes of the plotted lines.}

\item{margin_bottom, margin_left, margin_right, margin_top}{Numeric for the area around the plot. See \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{n_cols}{Numeric, indicating how many columns of smaller plots the final lineplot should have.}

\item{point_label_nudge}{Logical. If \code{TRUE}, the point labels will be nudged automatically by \link[ggrepel:geom_text_repel]{ggrepel::geom_text_repel} to try and avoid them overlapping with lines and/or points. Might interfere with other point-label settings, so use at your own risk. Alternatively, you could try to use \code{point_label_nudge_direction} to get some controll over the point labels.}

\item{point_label_size}{Numeric for the fontsize of the pointlabels.}

\item{point_label_nudge_direction}{Named list with contents of either "+" or "-". The names have to be the factorlevels of the grouping_var. For "+" the point lables will be printed above the point, for "-" below. If \code{NULL} is provided, the labels will be printed below the points for the lowest group, and above the points for all others.}

\item{point_label_nudge_x, point_label_nudge_y}{Numeric for the amount the pointlabel is nudged in x- or y-direction.}

\item{point_size}{Numeric for the size of plotted points.}

\item{split_plot}{Logical, indicating whether the different trends should be split or not.}

\item{split_plot_gap_width}{Numeric for the width of the gap in a split plot in npc.}

\item{y_axis}{Logical, indicating whether a y-axis should be plotted to the left of each row or not.}

\item{default_list}{Named list with predefined settings. Defaults to a list with all settings set to \code{0}.}
}
\value{
A named list with settings for the lineplots.
}
\description{
Set parameters for the lineplots.
}
\examples{
plotsettings(n_cols = 2, axis_x_label_centralize = 0.1)
}
