% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsettings_tablebarplot.R
\name{plotsettings_tablebarplot}
\alias{plotsettings_tablebarplot}
\title{Set parameters for the barplots.}
\usage{
plotsettings_tablebarplot(
  axis_x_lims = NULL,
  background_stripes_colour = NULL,
  bar_fill_colour = NULL,
  bar_frame_linetype = NULL,
  bar_pattern_fill_colour = NULL,
  bar_pattern_type = NULL,
  bar_sig_type = NULL,
  default_list = NULL
)
}
\arguments{
\item{axis_x_lims}{Numeric vector of length \code{2} for the x-axis limits. Will be set automatically, if \code{NULL}.}

\item{background_stripes_colour}{Named vector containing the two colours that should be used for the striped background.}

\item{bar_fill_colour}{Colour of the bar filling. Can be either one colour for all bars, or a named vector with the names of the groups specified in \code{bar_fill}.}

\item{bar_frame_linetype}{Named vector with the bar frame linetypes.}

\item{bar_pattern_fill_colour}{Named vector with the filling colours for the bar pattern. Names of the vector must be found in the column specified in \code{bar_pattern_fill}.}

\item{bar_pattern_type}{Named vector with the pattern types for the barpattern.}

\item{bar_sig_type}{Character string indicating whether levels of the grouping variable should be visualized by pattern fill ("pattern") or line type ("frame").}

\item{default_list}{Named list with predefined settings. Defaults to a list with all settings set to \code{0}.}
}
\value{
A named list with settings for the barplots.
}
\description{
Set parameters for the barplots.
}
\examples{
plotsettings_tablebarplot()

}
