% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_barplot.R
\name{prep_barplot}
\alias{prep_barplot}
\title{Prepare data for plotting a barplot.}
\usage{
prep_barplot(data, subGroups, sigNiveau)
}
\arguments{
\item{data}{Output from \code{}}

\item{subGroups}{Subgroups which should be plotted in the same plot tile. E.g. you can distinguish between adjusted and not adjusted means.}

\item{sigNiveau}{At which significance niveau do p-values become significant?}
}
\value{
Returns a \code{data.frame} which can be used as input for \code{plot_bar}.
}
\description{
Prepare data for plotting a barplot.
}
\examples{
#tbd
}
