% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_general.R
\name{prep_general}
\alias{prep_general}
\title{Prepare eatRep output for further data wrangling.}
\usage{
prep_general(data_clean, sig_niveau, BLs, groups)
}
\arguments{
\item{data_clean}{Input data.frame, that has already been cleaned with \code{\link[=clean_data]{clean_data()}}.}

\item{sig_niveau}{Numeric indicating the border, below which p-values will be considered significant. Defaults to \code{0.05}.}
}
\value{
List of data frames with different filters.
}
\description{
Prepare eatRep output for further data wrangling.
}
\examples{
# tbd
}
