% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_general.R
\name{prep_general}
\alias{prep_general}
\title{Prepare eatRep output for further data wrangling.}
\usage{
prep_general(data, grouping_var = "", competence)
}
\arguments{
\item{data}{eatRep output data.frame.}

\item{grouping_var}{Variable with subgroups.}

\item{competence}{Competence variable.}
}
\value{
List of data frames with different filters.
}
\description{
Prepare eatRep output for further data wrangling.
}
\examples{
# tbd
}
