% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_lineplot.R
\name{prep_lineplot}
\alias{prep_lineplot}
\title{Title}
\usage{
prep_lineplot(data, grouping_var, sig_niveau)
}
\arguments{
\item{data}{Data for line plot.}

\item{grouping_var}{Variable to group by.}

\item{sig_niveau}{Significance niveau.}
}
\value{
Data frame in long format.
}
\description{
Title
}
\examples{
#tbd
}
