% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_lineplot.R
\name{prep_lineplot}
\alias{prep_lineplot}
\title{Title}
\usage{
prep_lineplot(data, grouping_var, competence, sig_niveau = 0.05)
}
\arguments{
\item{data}{Trend data from eatRep.}

\item{grouping_var}{Grouping variable.}

\item{competence}{Competence area.}

\item{sig_niveau}{Significance niveau.}
}
\value{
List of data.frames needed for different plot-functions.
}
\description{
Title
}
\examples{
# tbd
}
