% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_lineplot.R
\name{prep_lineplot}
\alias{prep_lineplot}
\title{Plot a lineplot.}
\usage{
prep_lineplot(eatRep_dat, line_sig, parameter, years_lines, years_braces)
}
\arguments{
\item{eatRep_dat}{Object returned by \code{eatRep}.}
}
\value{
Data prepared for plotting the BT-lineplots.
}
\description{
Plot a lineplot.
}
\examples{
# tbd
}
