% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_no_trend.R
\name{prep_no_trend}
\alias{prep_no_trend}
\title{Prepare data for plotting a barplot.}
\usage{
prep_no_trend(data, grouping_var, columns, competence, sig_niveau)
}
\arguments{
\item{data}{Input data.frame stemming from \code{eatRep}.}

\item{grouping_var}{Character string containing the column in \code{data} that should be used to distinguish between subgroups.}

\item{columns}{Character string containing the column in \code{data} that contains the column names for the table_plot.}

\item{competence}{Character string containing the competence that should be plotted. Currently has to be found in \code{data$kb} (even though that should be made optional in the future).}

\item{sig_niveau}{Numeric indicating the border, below which p-values will be considered significant. Defaults to \code{0.05}.}
}
\value{
Returns a data.frame which can be used as input for plot_bar.
}
\description{
Prepare data for plotting a barplot.
}
\examples{
# tbd
}
