% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_points.R
\name{prep_points}
\alias{prep_points}
\title{Title}
\usage{
prep_points(data, competence, grouping_var)
}
\arguments{
\item{data}{Trend data.}

\item{competence}{Competence area.}

\item{grouping_var}{Variable that groups the cases.}
}
\value{
data.frame in long format.
}
\description{
Title
}
\examples{
# tbd
}
