% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tablebarplot.R
\name{prep_tablebarplot}
\alias{prep_tablebarplot}
\title{Prepare lineplot data.}
\usage{
prep_tablebarplot(
  eatRep_dat,
  subgroup_var,
  names_from = c("year", "comparison", "parameter"),
  par = "mean",
  facet_var = "TR_BUNDESLAND",
  total_group = "total",
  comparisons = NULL
)
}
\arguments{
\item{eatRep_dat}{Object returned by \code{eatRep}.}

\item{facet_var}{Character string of the name of the variable that should be used for faceting the plot. Defaults to \code{"TR_BUNDESLAND"}.}

\item{total_group}{Character string of the name of the total groups containing all other groups of the facet var. Defaults to \code{"total"}.}

\item{comparisons}{Character string that can be used to filter the needed comparions. This can drastically reduce the output of the prepared data. Defaults to \code{NULL}.}
}
\value{
Data prepared for plotting the BT-lineplots.
}
\description{
Prepare lineplot data.
}
\examples{
# tbd
}
