% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tablebarplot.R
\name{prep_tablebarplot}
\alias{prep_tablebarplot}
\title{Prepare tableplot data.}
\usage{
prep_tablebarplot(
  eatRep_dat,
  subgroup_var = NULL,
  parameter = "mean",
  facet_var = "TR_BUNDESLAND",
  total_facet = "total",
  sig_niveau = 0.05,
  total_subgroup = "total",
  parameter_to_colname = TRUE
)
}
\arguments{
\item{eatRep_dat}{Object returned by \code{eatRep}.}

\item{subgroup_var}{Character string of the column in \code{eatPlot_dat$group} or \code{eatPlot_dat$plain} containing the subgroup mapping. Each supgroup will receive it's own line. IF there are subgroups within the data, this needs to be set, otherwise the data preparation might fail.  Defaults to \code{NULL}.}

\item{parameter}{Character string. Contains the value in column \code{parameter} that is used for plotting the lines.  Defaults to \code{"mean"}.}

\item{facet_var}{Character string of the variable containing information on groups some of the comparisons are made against. This is needed to decosntruct comparisons like \code{crossDiff} into \code{crossDiff} and \code{crossDiffTotal} (so a crossDiff comparison against the total group). Name might be a bit confusing, but is the same as in \code{prep_lineplot}. Defaults to \code{TR_BUNDESLAND}.}

\item{total_facet}{Character string of the name of the total groups containing all other groups of the facet var. Defaults to \code{"total"}.}

\item{sig_niveau}{Numeric indicating the border below which p-values will be considered significant. Defaults to \code{0.05}.}

\item{total_subgroup}{Character string indicating the subgroup containing all other groups of the subgroup_var. Defaults to \code{"total"}.}

\item{parameter_to_colname}{Logical that indicates whether the parameter should be also pivoted into wide format, or the parameter column should be kept in the long format. Mainly set to \code{FALSE} for stacked barplots. Defaults to \code{TRUE}.}
}
\value{
Data prepared for plotting the BT-lineplots.
}
\description{
Prepare tableplot data.
}
\examples{
# tbd
}
