% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_trend.R
\name{prep_trend}
\alias{prep_trend}
\title{Title}
\usage{
prep_trend(data, sig_niveau = 0.05)
}
\arguments{
\item{data}{Data for line plot.}

\item{sig_niveau}{Significance niveau.}

\item{grouping_var}{Variable to group by.}
}
\value{
Data frame in long format, which includes the relevant trend variables.
}
\description{
Title
}
\examples{
#tbd
}
