% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_bundesland}
\alias{process_bundesland}
\title{Process the data in the Bundesland column.}
\usage{
process_bundesland(vec, linebreak = FALSE, total_group = "total")
}
\arguments{
\item{vec}{Character vector that contains Bunesl채nder that should be processed.}

\item{linebreak}{Logical. If \code{TRUE}, a linebreak is added after the '-'. Defaults to \code{FALSE}.}

\item{total_group}{Character string that should be replaced by \code{Deutschland}. Defaults to \code{"total"}.}
}
\value{
The eatPlot_data data.frame with the bundesl채nder column processed.
}
\description{
This changes \code{ue} to \code{체} and adds a \code{-} between bundesl채nder consisting of two words.
}
\examples{
# tbd
}
