% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Title}
\usage{
save_plot(
  p,
  filename,
  width = 160,
  height = 226.2,
  format = "PDF",
  scaling = 1
)
}
\arguments{
\item{p}{Plot that should be saved.}

\item{filename}{Filepath and -name for the new plot file.}

\item{width}{Width of the new plot file in mm. Defaults to \code{160}.}

\item{height}{Height of the new plot file in mm. Defaults to \code{226.2}}

\item{format}{Format of the new plot file. Currently, only PDF is needed.}

\item{scaling}{Proportional width and height changes, so the proportion between them is kept.}
}
\value{
A PDF file.
}
\description{
Title
}
\examples{
# tbd
}
