% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Wrapper for \code{\link[grDevices:pdf]{grDevices::pdf()}}.}
\usage{
save_plot(
  p,
  filename,
  width = 160,
  height = 226.2,
  colormodel = "cmyk",
  scaling = 1,
  ...
)
}
\arguments{
\item{p}{Plot that should be saved.}

\item{filename}{Filepath and -name for the new plot file.}

\item{width}{Width of the new plot file in mm. Defaults to \code{160}.}

\item{height}{Height of the new plot file in mm. Defaults to \code{226.2}}

\item{colormodel}{Colormodel for saving the plot. See \code{\link[grDevices:pdf]{grDevices::pdf()}} for more information. Defaults to "cmyk".}

\item{scaling}{Scaling parameter for changing height in widht relationally.}

\item{...}{Further arguments passed to \code{\link[grDevices:pdf]{grDevices::pdf()}}.}
}
\value{
A PDF file.
}
\description{
Wrapper for \code{\link[grDevices:pdf]{grDevices::pdf()}}.
}
\examples{
# tbd
}
