% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Wrapper for \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
\usage{
save_plot(p, filename, width = 160, height = 226.2, device = cairo_pdf, ...)
}
\arguments{
\item{p}{Plot that should be saved.}

\item{filename}{Filepath and -name for the new plot file.}

\item{width}{Width of the new plot file in mm. Defaults to \code{160}.}

\item{height}{Height of the new plot file in mm. Defaults to \code{226.2}}

\item{device}{Device for saving the plot. Should be \code{cairo_pdf} for pdfs, so font is displayed correctly.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}

\item{format}{Format of the new plot file. Currently, only PDF is needed.}
}
\value{
A PDF file.
}
\description{
Wrapper for \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.
}
\examples{
# tbd
}
