% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plot_coords.R
\name{set_plot_coords}
\alias{set_plot_coords}
\title{Set the x- and y-coordinates for a plot.}
\usage{
set_plot_coords(
  plot_dat,
  plot_lims = plot_lims,
  plot_settings = plotsettings_lineplot()
)
}
\arguments{
\item{plot_dat}{Input is a list prepared by \code{prep_plot()}.}
}
\value{
\link{ggplot2} object.
}
\description{
Set the x- and y-coordinates for a plot.
}
\examples{
# tbd
}
