% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_plot_coords.R
\name{set_plot_coords}
\alias{set_plot_coords}
\title{Set the x- and y-coordinates for a plot.}
\usage{
set_plot_coords(
  plot_dat,
  x_range,
  y_range,
  plot_settings = plotsettings_lineplot()
)
}
\arguments{
\item{plot_dat}{Input is a list prepared by \code{\link[=prep_plot]{prep_plot()}}.`}

\item{x_range}{Numeric vector with two elements, indicating the min and max for the x axis.}

\item{y_range}{Numeric vector with two elements, indicating the min and max for the y axis.}

\item{plot_settings}{Named list constructed with \code{plotsettings_lineplot()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plotsettings_lineplot()} for an overview.}
}
\value{
\link{ggplot2} object.
}
\description{
Set the x- and y-coordinates for a plot.
}
\examples{
# tbd
}
