% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_plot.R
\name{settings_lineplot}
\alias{settings_lineplot}
\title{Set theme, colours, pointshapes and linetypes for lineplot.}
\usage{
settings_lineplot(plot_settings = plotsettings())
}
\arguments{
\item{plot_settings}{Named list constructed with \code{plot_settings()}. Defaults to a list with all settings set to \code{0}. There are several predefined lists with optimized settings for different plots. See \code{plot_settings()} for an overview.}
}
\value{
ggplot2 plot settings
}
\description{
Set theme, colours, pointshapes and linetypes for lineplot.
}
\examples{
#tbd
}
