% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings_plot.R
\name{settings_lineplot}
\alias{settings_lineplot}
\title{Set theme, colours, pointshapes and linetypes for lineplot.}
\usage{
settings_lineplot(plot_settings = plotsettings())
}
\value{
ggplot2 plot settings
}
\description{
Set theme, colours, pointshapes and linetypes for lineplot.
}
\examples{
#tbd
}
