% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_y_axis}
\alias{theme_y_axis}
\title{Theme for a y-axis plot.}
\usage{
theme_y_axis()
}
\value{
\link{ggplot2} theme, that can be used for a plotted y-axis.
}
\description{
Theme for a y-axis plot.
}
\examples{
# tbd
}
